added routes and user login/register UI

+++ src/hooks/useTodo.ts
@@ -1,6 +1,9 @@
 import { useState, useEffect, useRef } from 'react';
+import { useLoaderData } from 'react-router-dom';
 import { TodoEntity } from '../../src/common/interfaces/ITodoEntity';
 import { IHttpClient } from '../common/interfaces/IHttpClient';
+import { AuthToken } from '../common/services/AuthToken'
+import { useNavigate } from 'react-router-dom'
 
 export default function useTodo(httpClient: IHttpClient, url: string) {
     const [todos, setTodos] = useState<TodoEntity[]>([]);
@@ -8,16 +11,20 @@ export default function useTodo(httpClient: IHttpClient, url: string) {
     const [newTodo, setNewTodo] = useState<string>('');
     const [addTodoError, setAddTodoError] = useState<boolean>(false);
     const [deleteAllModal, setDeleteAllModal] = useState<boolean>(false)
+    const [user, setUser] = useState<String>('')
+
+    const navigate = useNavigate()
+    const initialData: any = useLoaderData()
 
     useEffect(() => {
-        const getAllTodos = async () => {
-            const data = await httpClient.sendRequest(url, { method: 'GET' });
-            if (data.response) {
-                setTodos(data.response);
-            }
-        };
 
-        getAllTodos();
+        if(initialData){
+            setTodos(initialData.todos)
+            setUser(initialData.user)
+        } else {
+            navigate("/login")
+        }
+
     }, []);
 
     useEffect(() => {
@@ -36,29 +43,54 @@ export default function useTodo(httpClient: IHttpClient, url: string) {
     const lastTodoRef = useRef<null | HTMLDivElement>(null);
 
     const deleteTodo = async (id: string) => {
+        const token = new AuthToken('user').getToken()
+
+        if(!token){
+            navigate('/login')
+        }
+
         const data = await httpClient.sendRequest(`${url}/${id}`, {
             method: 'DELETE',
-        });
+            headers: { 'Authorization': token }
+        })
+       
         if (data.response) {
             setTodos((prevTodos) => prevTodos.filter((todo) => todo.id !== id));
         }
     };
 
     const deleteAllTodos = async () => {
-        const data = await httpClient.sendRequest(`${url}/deleteAll`, { method: "DELETE"} )
+        const token = new AuthToken('user').getToken()
+        
+        
+        if(!token){
+            navigate('/login')
+        }
+
+        const data = await httpClient.sendRequest(`${url}/deleteAll`, { method: "DELETE", headers: { 'Authorization': token }})
+        .catch(err => console.log(err))
         
         if(data.response){
             setTodos([])
         }
+
         setDeleteAllModal(false)
     }
 
     const checkTodo = async (id: string, completed: boolean) => {
+        const token = new AuthToken('user').getToken()
+        
+        
+        if(!token){
+            navigate('/login')
+        }
+
         const data = await httpClient.sendRequest(`${url}/${id}`, {
             method: 'PUT',
-            headers: { 'Content-Type': 'application/json' },
+            headers: { 'Content-Type': 'application/json', 'Authorization': token },
             body: JSON.stringify({ completed: !completed }),
-        });
+        }).catch(err => console.log(err))
+
         if (data.response) {
             setTodos((prevTodos) =>
                 prevTodos.map((todo) =>
@@ -73,11 +105,18 @@ export default function useTodo(httpClient: IHttpClient, url: string) {
             return setAddTodoError(true);
         }
 
+        const token = new AuthToken('user').getToken()
+
+        if(!token){
+            navigate('/login')
+        }
+
         const data = await httpClient.sendRequest(url, {
             method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
+            headers: { 'Content-Type': 'application/json', 'Authorization': token },
             body: JSON.stringify({ text: todo }),
-        });
+        })
+        .catch(err => console.log(err))
 
         setNewTodo('');
         setTodoWindow(false);
@@ -93,6 +132,7 @@ export default function useTodo(httpClient: IHttpClient, url: string) {
 
     return {
         todos,
+        user,
         todoWindow,
         newTodo,
         setNewTodo,
